pyproject.toml
[project]
name = "bootagent"
version = "0.1.0"
description = "AI-powered coding assistant using Google Gemini function calling"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "python-dotenv==1.1.0",
]
